128:Lazy Loading
Doing lazy loading by passing pages as strings, increases the performance of the application because the pages will load when the user asks for them.
If not pass as strings, then you have to import the page in the page that is asking for it, in addtion to the entery componenet and the app module. Additional work that reduces the performance of the application