135. Component Communication with EventEmitters
Since we have separate components than the page component, event communcation between the component and the page component is shown here. for the login, register and error messages that happen. This part seems to be a bit complicated. Max might have a easier solution. There are many ways to do this.