138. Creating an Edit Profile Form
Once you register you are taken to a edit profile form. The HTML and login for this EditprofileForm is in a different component. We just simply use the selector in the page to show the content in the page from the component. All components are exported in component.module.ts and imorted inside module of the page