140. Login Form/Flow Improvements (WORKING)
This section does the following:
if profile doesn't exists, you are routed to the edit profile from login, if it does exists you are routed to  tabs page directly. Also once profile save was complete, you are routed to the Tabs page, if the save doesn't complete, it logs to the console that you are not  logged in. although routing can be done in the component, but he wanted to show the event emitters and output so component (child) passes data  to the page component (parent( using eventemitters and output.