144. Edit Profile as an Authenticated User (very good component communication example)
Complete usage of event emitters, input, output
After these changes user can click on the button that is created on the profile page after creating the account to edit the already created profile.

I also had to change back the security becuase the premission for accessing data was denied. (need to work on this later):
{
  "rules": {
    ".read": "auth != null",
    ".write": "auth != null"
  }
}