154. Displaying Channel List (IMPORTANT: previous 154 added  should be  153. Adding Chat Channels
IMPRTANT:
The commit that was before this one should have been   153. Adding Chat Channels where we added an alert using a plus sign on top right of the  page to display an alert where user could add a channel.
This commit 154. Displaying Channel List, displays the channel that are available on the channel page itself